refactor(22.5): separate token validation in a middleware
