+++ back-end/bloco-22-introducao-ao-desenvolvimento-web-com-nodejs/dia-5-express-middlewares/exercise-2/index.js
@@ -1,25 +1,12 @@
 const express = require('express');
 const axios = require('axios').default;
+const tokenValidationMiddleware = require('./tokenValidationMiddleware');
 
 const app = express();
 
 app.use(express.json());
 
-function hasOnlyLettersAndNumbers(string) {
-  const validationRegex = /^[A-Z0-9]+$/i;
-  return validationRegex.test(string);
-}
-
-function isTokenValid(token) {
-  return token && token.length === 12 && hasOnlyLettersAndNumbers(token);
-}
-
-app.get('/btc/price', async (req, res) => {
-  const token = req.headers.authorization;
-
-  if (!isTokenValid(token)) {
-    return res.status(401).json({ message: 'invalid token' });
-  }
+app.get('/btc/price', tokenValidationMiddleware, async (_req, res) => {
   const response = await axios.get('https://api.coindesk.com/v1/bpi/currentprice/BTC.json');
   res.status(200).json(response.data);
 });
